# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PrizePicks Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams
      parameters:
        - { name: sport, in: query, schema: { type: string } }
      responses:
        "200":
          description: Team list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Team" }
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create team
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }

  # ─── GAMES ────────────────────────────────────────────────────────
...
  /slate:
    get:
      operationId: getSlate
      tags: [propScores]
      summary: Get enriched slate — props with scores, projections, player/team/game context
      parameters:
        - { name: sport, in: query, schema: { type: string } }
        - { name: statType, in: query, schema: { type: string } }
        - { name: actionTag, in: query, schema: { type: string } }
        - { name: lineType, in: query, schema: { type: string } }
        - { name: teamId, in: query, schema: { type: integer } }
        - { name: gameId, in: query, schema: { type: integer } }
        - { name: minEdgeScore, in: query, schema: { type: number } }
        - { name: maxRiskScore, in: query, schema: { type: number } }
        - { name: watchlistOnly, in: query, schema: { type: boolean } }
        - { name: date, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Enriched slate rows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SlateRow" }
  /slate/{ppLineId}:
    get:
      operationId: getSlateRow
      tags: [propScores]
      summary: Get full prop detail for one line
      parameters:
        - { name: ppLineId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Full prop detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PropDetail" }

  # ─── ENTRIES ──────────────────────────────────────────────────────
...
  /watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: List watchlist items
      parameters:
        - { name: playerId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WatchlistItem" }
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchlistItemInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistItem" }
  /watchlist/{id}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove from watchlist
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
    patch:
      operationId: updateWatchlistItem
      tags: [watchlist]
      summary: Update watchlist item
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchlistItemUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistItem" }

  # ─── ALERTS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id: { type: integer }
        sport: { type: string }
        fullName: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        teamId: { type: ["integer", "null"] }
        position: { type: ["string", "null"] }
        status: { type: string }
        externalIds: {}
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, sport, fullName, firstName, lastName, status, createdAt, updatedAt]
...
    Team:
      type: object
      properties:
        id: { type: integer }
        sport: { type: string }
        name: { type: string }
        abbreviation: { type: string }
        city: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, sport, name, abbreviation, createdAt, updatedAt]
    TeamInput:
      type: object
      properties:
        sport: { type: string }
        name: { type: string }
        abbreviation: { type: string }
        city: { type: ["string", "null"] }
      required: [sport, name, abbreviation]

    # ─── GAME ───────────────────────────────────────────────────────
...
    PpLine:
      type: object
      properties:
        id: { type: integer }
        playerId: { type: integer }
        gameId: { type: ["integer", "null"] }
        statType: { type: string }
        directionalityType: { type: string }
        lineValue: { type: number }
        lineType: { type: string }
        payoutModifier: {}
        openedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, playerId, statType, directionalityType, lineValue, lineType, isActive, createdAt, updatedAt, openedAt]
...
    PpLineHistory:
      type: object
      properties:
        id: { type: integer }
        ppLineId: { type: integer }
        lineValue: { type: number }
        lineType: { type: string }
        capturedAt: { type: string, format: date-time }
        metadata: {}
      required: [id, ppLineId, lineValue, lineType, capturedAt]

    # ─── EXTERNAL LINE ──────────────────────────────────────────────
    ExternalLine:
      type: object
      properties:
        id: { type: integer }
        playerId: { type: integer }
        gameId: { type: ["integer", "null"] }
        statType: { type: string }
        bookName: { type: string }
        overLine: { type: number }
        overOdds: { type: ["integer", "null"] }
        underLine: { type: number }
        underOdds: { type: ["integer", "null"] }
        noVigOverProb: { type: ["number", "null"] }
        noVigUnderProb: { type: ["number", "null"] }
        pulledAt: { type: string, format: date-time }
        metadata: {}
      required: [id, playerId, statType, bookName, overLine, underLine, pulledAt]

    # ─── PROJECTION ─────────────────────────────────────────────────
...
    Injury:
      type: object
      properties:
        id: { type: integer }
        playerId: { type: integer }
        playerName: { type: ["string", "null"] }
        playerTeam: { type: ["string", "null"] }
        gameId: { type: ["integer", "null"] }
        sport: { type: string }
        status: { type: string }
        note: { type: string }
        source: { type: string }
        reportedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
      required: [id, playerId, sport, status, note, source, reportedAt, createdAt]
...
    SlateRow:
      type: object
      properties:
        ppLineId: { type: integer }
        playerId: { type: integer }
        playerName: { type: string }
        teamAbbr: { type: ["string", "null"] }
        opponentAbbr: { type: ["string", "null"] }
        sport: { type: string }
        startTime: { type: ["string", "null"] }
        statType: { type: string }
        lineValue: { type: number }
        lineType: { type: string }
        directionalityType: { type: string }
        yourProjection: { type: ["number", "null"] }
        projectionGap: { type: ["number", "null"] }
        edgeScore: { type: ["number", "null"] }
        stabilityScore: { type: ["number", "null"] }
        marketSupportScore: { type: ["number", "null"] }
        riskScore: { type: ["number", "null"] }
        finalScore: { type: ["number", "null"] }
        actionTag: { type: ["string", "null"] }
        isWatched: { type: boolean }
        updatedAt: { type: string, format: date-time }
      required: [ppLineId, playerId, playerName, sport, statType, lineValue, lineType, directionalityType, isWatched, updatedAt]

    # ─── PROP DETAIL ────────────────────────────────────────────────
    PropDetail:
      type: object
      properties:
        ppLine: { $ref: "#/components/schemas/PpLine" }
        player: { $ref: "#/components/schemas/Player" }
        game: {}
        lineHistory: { type: array, items: { $ref: "#/components/schemas/PpLineHistory" } }
        projection: {}
        externalLines: { type: array, items: { $ref: "#/components/schemas/ExternalLine" } }
        propScore: {}
        injuries: { type: array, items: { $ref: "#/components/schemas/Injury" } }
        lineupConfirmation: {}
        isWatched: { type: boolean }
        watchlistId: { type: ["integer", "null"] }
      required: [ppLine, player, lineHistory, externalLines, injuries, isWatched]

    # ─── ENTRY ──────────────────────────────────────────────────────
...
    WatchlistItem:
      type: object
      properties:
        id: { type: integer }
        playerId: { type: integer }
        playerName: { type: ["string", "null"] }
        gameId: { type: ["integer", "null"] }
        statType: { type: string }
        directionPreference: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, playerId, statType, createdAt]
    WatchlistItemInput:
      type: object
      properties:
        playerId: { type: integer }
        gameId: { type: ["integer", "null"] }
        statType: { type: string }
        directionPreference: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
      required: [playerId, statType]
    WatchlistItemUpdate:
      type: object
      properties:
        directionPreference: { type: ["string", "null"] }
        note: { type: ["string", "null"] }

    # ─── ALERT ──────────────────────────────────────────────────────
...